fix(replace): replace all code
